requirements.txt
aiohttp
python-dotenv
fastapi
twilio
uvicorn
pipecat
openai
websocket-client
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-openai
livekit-plugins-silero
livekit-protocol
livekit-plugins-turn-detector
resend
uvicorn